requirements.txt
streamlit==1.29.0
Pillow==10.1.0
tweepy==4.14.0
requests==2.31.0
